Выносит общую логику обновления в smart.js
